+++ lib/custom_widgets/markdown_config.dart
@@ -64,12 +64,12 @@ typedef EmojiBuilder = Widget Function(BuildContext context, String emoji);
 /// The [GptMarkdownConfig] class is used to configure the GPT Markdown component.
 /// It takes a [style] parameter to set the style of the text,
 /// a [textDirection] parameter to set the direction of the text,
-/// and an optional [onLinkTab] parameter to handle link clicks.
+/// and an optional [onLinkTap] parameter to handle link clicks.
 class GptMarkdownConfig {
   const GptMarkdownConfig({
     this.style,
     this.textDirection = TextDirection.ltr,
-    this.onLinkTab,
+    this.onLinkTap,
     this.textAlign,
     this.textScaler,
     this.latexWorkaround,
@@ -102,7 +102,7 @@ class GptMarkdownConfig {
   final TextScaler? textScaler;
 
   /// The callback function to handle link clicks.
-  final void Function(String url, String title)? onLinkTab;
+  final void Function(String url, String title)? onLinkTap;
 
   /// The LaTeX workaround.
   final String Function(String tex)? latexWorkaround;
@@ -153,7 +153,7 @@ class GptMarkdownConfig {
   GptMarkdownConfig copyWith({
     TextStyle? style,
     TextDirection? textDirection,
-    final void Function(String url, String title)? onLinkTab,
+    final void Function(String url, String title)? onLinkTap,
     final TextAlign? textAlign,
     final TextScaler? textScaler,
     final String Function(String tex)? latexWorkaround,
@@ -175,7 +175,7 @@ class GptMarkdownConfig {
     return GptMarkdownConfig(
       style: style ?? this.style,
       textDirection: textDirection ?? this.textDirection,
-      onLinkTab: onLinkTab ?? this.onLinkTab,
+      onLinkTap: onLinkTap ?? this.onLinkTap,
       textAlign: textAlign ?? this.textAlign,
       textScaler: textScaler ?? this.textScaler,
       latexWorkaround: latexWorkaround ?? this.latexWorkaround,
@@ -206,5 +206,28 @@ class GptMarkdownConfig {
       maxLines: maxLines,
       overflow: overflow,
     );
+  }
+
+  /// A method to check if the configuration is the same.
+  bool isSame(GptMarkdownConfig other) {
+    return style == other.style &&
+        textAlign == other.textAlign &&
+        textScaler == other.textScaler &&
+        maxLines == other.maxLines &&
+        overflow == other.overflow &&
+        followLinkColor == other.followLinkColor &&
+        // latexWorkaround == other.latexWorkaround &&
+        // components == other.components &&
+        // inlineComponents == other.inlineComponents &&
+        // latexBuilder == other.latexBuilder &&
+        // sourceTagBuilder == other.sourceTagBuilder &&
+        // codeBuilder == other.codeBuilder &&
+        // orderedListBuilder == other.orderedListBuilder &&
+        // unOrderedListBuilder == other.unOrderedListBuilder &&
+        // linkBuilder == other.linkBuilder &&
+        // imageBuilder == other.imageBuilder &&
+        // highlightBuilder == other.highlightBuilder &&
+        // onLinkTap == other.onLinkTap &&
+        textDirection == other.textDirection;
   }
 }