+++ lib/gpt_markdown.dart
@@ -30,7 +30,7 @@ class GptMarkdown extends StatelessWidget {
     this.textAlign,
     this.imageBuilder,
     this.textScaler,
-    this.onLinkTab,
+    this.onLinkTap,
     this.latexBuilder,
     this.codeBuilder,
     this.sourceTagBuilder,
@@ -43,6 +43,7 @@ class GptMarkdown extends StatelessWidget {
     this.unOrderedListBuilder,
     this.components,
     this.inlineComponents,
+    this.useDollarSignsForLatex = false,
   });
 
   /// The direction of the text.
@@ -61,7 +62,7 @@ class GptMarkdown extends StatelessWidget {
   final TextScaler? textScaler;
 
   /// The callback function to handle link clicks.
-  final void Function(String url, String title)? onLinkTab;
+  final void Function(String url, String title)? onLinkTap;
 
   /// The LaTeX workaround.
   final String Function(String tex)? latexWorkaround;
@@ -99,6 +100,9 @@ class GptMarkdown extends StatelessWidget {
 
   /// The unordered list builder.
   final UnOrderedListBuilder? unOrderedListBuilder;
+
+  /// Whether to use dollar signs for LaTeX.
+  final bool useDollarSignsForLatex;
 
   /// The list of components.
   ///  ```dart
@@ -158,31 +162,34 @@ class GptMarkdown extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     String tex = data.trim();
-    tex = tex.replaceAllMapped(
-      RegExp(r"(?<!\\)\$\$(.*?)(?<!\\)\$\$", dotAll: true),
-      (match) => "\\[${match[1] ?? ""}\\]",
-    );
-    if (!tex.contains(r"\(")) {
+    if (useDollarSignsForLatex) {
       tex = tex.replaceAllMapped(
-        RegExp(r"(?<!\\)\$(.*?)(?<!\\)\$"),
-        (match) => "\\(${match[1] ?? ""}\\)",
+        RegExp(r"(?<!\\)\$\$(.*?)(?<!\\)\$\$", dotAll: true),
+        (match) => "\\[${match[1] ?? ""}\\]",
       );
-      tex = tex.splitMapJoin(
-        RegExp(r"\[.*?\]|\(.*?\)"),
-        onNonMatch: (p0) {
-          return p0.replaceAll("\\\$", "\$");
-        },
-      );
+      if (!tex.contains(r"\(")) {
+        tex = tex.replaceAllMapped(
+          RegExp(r"(?<!\\)\$(.*?)(?<!\\)\$"),
+          (match) => "\\(${match[1] ?? ""}\\)",
+        );
+        tex = tex.splitMapJoin(
+          RegExp(r"\[.*?\]|\(.*?\)"),
+          onNonMatch: (p0) {
+            return p0.replaceAll("\\\$", "\$");
+          },
+        );
+      }
     }
     // tex = _removeExtraLinesInsideBlockLatex(tex);
     return ClipRRect(
       child: MdWidget(
+        context,
         tex,
         true,
         config: GptMarkdownConfig(
           textDirection: textDirection,
           style: style,
-          onLinkTab: onLinkTab,
+          onLinkTap: onLinkTap,
           textAlign: textAlign,
           textScaler: textScaler,
           followLinkColor: followLinkColor,